+++ cmd/command/cd/cd_stacks.go
@@ -2,6 +2,7 @@ package cd
 
 import (
 	"fmt"
+
 	"github.com/pluralsh/plural-cli/pkg/common"
 
 	"github.com/AlecAivazis/survey/v2"
