+++ pkg/client/plural.go
@@ -2,12 +2,13 @@ package client
 
 import (
 	"fmt"
+	"os"
+
 	"github.com/pluralsh/plural-cli/pkg/common"
 	"github.com/pluralsh/plural-cli/pkg/crypto"
 	"github.com/pluralsh/plural-cli/pkg/scm"
 	"github.com/pluralsh/plural-cli/pkg/wkspace"
 	"github.com/urfave/cli"
-	"os"
 
 	"github.com/pluralsh/plural-cli/pkg/api"
 	"github.com/pluralsh/plural-cli/pkg/config"
